feat: 배팅 금액 입력
