+++ README.md
@@ -157,7 +157,7 @@ jason: -20000
 ### 기능 구현 목록
 
 #### 인풋
-- [ ] 플레이어의 배팅 금액을 입력받는다.
+- [x] 플레이어의 배팅 금액을 입력받는다.
   - [ ] 배팅 금액이 숫자인지 검증한다.
   - [ ] 배팅 금액이 양수인지 검증한다.
