+++ src/main/java/blackjack/dto/response/CardCountingResult.java
@@ -1,24 +1,30 @@
-package blackjack.dto;
+package blackjack.dto.response;
 
 import blackjack.domain.card.Card;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class PlayerCardResult {
-    private final String playerName;
+public class CardCountingResult {
+    private final String name;
     private final List<String> cards;
+    private final int count;
 
-    public PlayerCardResult(String playerName, List<Card> cards) {
-        this.playerName = playerName;
+    public CardCountingResult(String playerName, List<Card> cards, int count) {
+        this.name = playerName;
         this.cards = toCardNames(cards);
+        this.count = count;
     }
 
-    public String getPlayerName() {
-        return playerName;
+    public String getName() {
+        return name;
     }
 
     public List<String> getCards() {
         return cards;
+    }
+
+    public int getCount() {
+        return count;
     }
 
     private List<String> toCardNames(List<Card> cards) {
