+++ src/main/java/blackjack/dto/response/PlayerGameResult.java
@@ -1,4 +1,4 @@
-package blackjack.dto;
+package blackjack.dto.response;
 
 import blackjack.domain.PlayerOutcome;
